SweetAlert aplicadas

+++ src/app/components/profile-card/profile-card.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, inject } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { UserInfo } from 'src/app/interfaces/user-info.interface';
 import { ProfilesService } from 'src/app/services/profiles.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-profile-card',
@@ -15,19 +16,29 @@ export class ProfileCardComponent {
   profileService = inject(ProfilesService);
   router = inject(Router);
 
-
-  // Crearemos la función deletePost
-  oneProfile!: UserInfo | any;
-
   async deleteProfile(_id: string): Promise<void> {
-    alert("¿Estás seguro de querer borrar este perfil?");
-    let response = await this.profileService.delete(_id);
-    if (response) {
-      alert("Perfil borrado correctamente");
-      this.router.navigate(['/home']);
-    } else {
-      alert("Algo ha fallado")
-      console.error("No se pudo borrar el perfil.")
+    const confirmDelete = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás deshacer esta acción!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, borrarlo',
+      cancelButtonText: 'Cancelar',
+    });
+  
+    if (confirmDelete.isConfirmed) {
+      let response = await this.profileService.delete(_id);
+      if (response) {
+        Swal.fire('Borrado', 'El perfil ha sido eliminado.', 'success');
+        this.router.navigate(['/home']);
+      } else {
+        Swal.fire('Error', 'Algo ha fallado', 'error');
+        console.error('No se pudo borrar el perfil.');
+      }
     }
   }
 }
+  
+  
+  
+
